refactor(guardian): remove dead forward.ts, extract session-target module
forward.ts was largely dead: only resolveSessionTarget was imported (by
proxy.ts). The session cache, per-key locks (withSessionLock), prune timer,
findExistingSessionId, askAssistant, clearAssistantSessions, shouldClearSession,
sessionCacheSize, and the exported SESSION_TTL_MS had no remaining consumers
(the live assistant path is the /oc proxy in proxy.ts; assistant-client.ts is
still used by moderation.ts, so it is retained).

Move the one live export (resolveSessionTarget + SessionTarget type) into a
small session-target.ts module and delete forward.ts. Behavior-preserving:
resolveSessionTarget logic is unchanged.
