--- packages/guardian/src/forward.ts
@@ -1,202 +1,0 @@
-/**
- * Assistant forwarding — session management and message routing.
- *
- * Manages session lifecycle (create, reuse, clear) with caching and
- * per-key locking to avoid duplicate session creation under concurrent
- * requests. Uses the shared assistant HTTP client (assistant-client.ts).
- */
-
-import {
-  createSession,
-  deleteSession,
-  listSessions,
-  sendMessage,
-} from './assistant-client.ts';
-import type { AssistantClientOptions } from './assistant-client.ts';
-import { asRecord } from './utils.ts';
-
-// ── Config ──────────────────────────────────────────────────────────────
-
-const ASSISTANT_URL = Bun.env.OP_ASSISTANT_URL ?? "http://assistant:4096";
-const MESSAGE_TIMEOUT = Number(Bun.env.OPENCODE_TIMEOUT_MS ?? 0);
-const SESSION_TTL_MS = Number(Bun.env.GUARDIAN_SESSION_TTL_MS ?? 15 * 60_000);
-const SESSION_KEY_MAX_LENGTH = 256;
-
-const CLIENT_OPTS: AssistantClientOptions = Object.freeze({
-  baseUrl: ASSISTANT_URL,
-  messageTimeoutMs: MESSAGE_TIMEOUT,
-});
-
-// ── Session cache ───────────────────────────────────────────────────────
-
-const sessionCache = new Map<string, { sessionId: string; lastUsed: number }>();
-const sessionLocks = new Map<string, Promise<unknown>>();
-
-// Periodic cleanup of expired sessions + hard cap at 10,000
-// unref() so the timer doesn't keep the event loop alive (cleaner testing + shutdown).
-const pruneSessionTimer = setInterval(() => {
-  const now = Date.now();
-  for (const [key, entry] of sessionCache) {
-    if (now - entry.lastUsed > SESSION_TTL_MS) sessionCache.delete(key);
-  }
-
-  // Hard cap: if still over 10,000 after pruning expired, delete oldest entries first
-  if (sessionCache.size > 10_000) {
-    const sorted = [...sessionCache.entries()].sort((a, b) => a[1].lastUsed - b[1].lastUsed);
-    const toRemove = sorted.slice(0, sorted.length - 10_000);
-    for (const [k] of toRemove) sessionCache.delete(k);
-  }
-}, 5 * 60_000);
-pruneSessionTimer.unref();
-
-// ── Session title cache (for reattaching to existing assistant sessions) ──
-
-const SESSION_LIST_CACHE_TTL_MS = 60_000;
-const sessionTitleCache = new Map<string, string>();
-let sessionListCacheLastLoaded = 0;
-
-// ── Types ───────────────────────────────────────────────────────────────
-
-export type SessionTarget = {
-  cacheKey: string;
-  sessionKey: string;
-  title: string;
-};
-
-// ── Public API ──────────────────────────────────────────────────────────
-
-export function resolveSessionTarget(userId: string, portalId: string, metadata: unknown): SessionTarget {
-  const meta = asRecord(metadata);
-  const metadataSessionKey = typeof meta?.sessionKey === "string"
-    ? meta.sessionKey.trim()
-    : "";
-  const sessionKey = metadataSessionKey && metadataSessionKey.length <= SESSION_KEY_MAX_LENGTH
-    ? metadataSessionKey
-    : userId;
-
-  return {
-    cacheKey: `${portalId}:${sessionKey}`,
-    sessionKey,
-    title: `${portalId}/${sessionKey}`,
-  };
-}
-
-export function shouldClearSession(metadata: unknown): boolean {
-  return asRecord(metadata)?.clearSession === true;
-}
-
-export async function askAssistant(
-  message: string,
-  sessionTarget: SessionTarget,
-): Promise<{ answer: string; sessionId: string }> {
-  return withSessionLock(sessionTarget.cacheKey, async () => {
-    const cacheKey = sessionTarget.cacheKey;
-    const cached = sessionCache.get(cacheKey);
-
-    if (cached && Date.now() - cached.lastUsed < SESSION_TTL_MS) {
-      try {
-        const answer = await sendMessage(CLIENT_OPTS, cached.sessionId, message);
-        cached.lastUsed = Date.now();
-        sessionTitleCache.set(sessionTarget.title, cached.sessionId);
-        return { answer, sessionId: cached.sessionId };
-      } catch {
-        sessionCache.delete(cacheKey);
-        sessionTitleCache.delete(sessionTarget.title);
-        sessionListCacheLastLoaded = 0;
-      }
-    }
-
-    const existingSessionId = await findExistingSessionId(sessionTarget);
-    if (existingSessionId) {
-      try {
-        const answer = await sendMessage(CLIENT_OPTS, existingSessionId, message);
-        sessionCache.set(cacheKey, { sessionId: existingSessionId, lastUsed: Date.now() });
-        sessionTitleCache.set(sessionTarget.title, existingSessionId);
-        return { answer, sessionId: existingSessionId };
-      } catch {
-        sessionCache.delete(cacheKey);
-        sessionTitleCache.delete(sessionTarget.title);
-        sessionListCacheLastLoaded = 0;
-      }
-    }
-
-    const sessionId = await createSession(CLIENT_OPTS, sessionTarget.title);
-    const answer = await sendMessage(CLIENT_OPTS, sessionId, message);
-    sessionCache.set(cacheKey, { sessionId, lastUsed: Date.now() });
-    sessionTitleCache.set(sessionTarget.title, sessionId);
-    return { answer, sessionId };
-  });
-}
-
-export async function clearAssistantSessions(sessionTarget: SessionTarget): Promise<void> {
-  await withSessionLock(sessionTarget.cacheKey, async () => {
-    sessionCache.delete(sessionTarget.cacheKey);
-    sessionTitleCache.delete(sessionTarget.title);
-    // Force the next findExistingSessionId to re-fetch the session list
-    sessionListCacheLastLoaded = 0;
-
-    const sessions = await listSessions(CLIENT_OPTS);
-    const matching = sessions.filter((session) => session.title === sessionTarget.title);
-
-    for (const session of matching) {
-      try {
-        await deleteSession(CLIENT_OPTS, session.id);
-      } catch {
-        // best-effort cleanup; cache removal already ensures a fresh mapping next turn
-      }
-    }
-  });
-}
-
-/** Expose active session count for the /stats endpoint. */
-export function sessionCacheSize(): number {
-  return sessionCache.size;
-}
-
-/** Expose session TTL for the /stats endpoint. */
-export { SESSION_TTL_MS };
-
-// ── Internal helpers ────────────────────────────────────────────────────
-
-async function withSessionLock<T>(cacheKey: string, fn: () => Promise<T>): Promise<T> {
-  const previous = sessionLocks.get(cacheKey) ?? Promise.resolve();
-
-  let release = () => {};
-  const current = new Promise<void>((resolve) => {
-    release = resolve;
-  });
-  const chain = current.catch(() => {});
-  sessionLocks.set(cacheKey, chain);
-
-  await previous.catch(() => {});
-
-  try {
-    return await fn();
-  } finally {
-    release();
-    if (sessionLocks.get(cacheKey) === chain) {
-      sessionLocks.delete(cacheKey);
-    }
-  }
-}
-
-async function findExistingSessionId(sessionTarget: SessionTarget): Promise<string | null> {
-  const now = Date.now();
-  const cachedId = sessionTitleCache.get(sessionTarget.title);
-  // Fresh cache hit: return immediately. Either a stale TTL or a miss (which could
-  // be an externally-created session) triggers a refresh.
-  if (cachedId && now - sessionListCacheLastLoaded < SESSION_LIST_CACHE_TTL_MS) {
-    return cachedId;
-  }
-
-  const sessions = await listSessions(CLIENT_OPTS);
-  sessionTitleCache.clear();
-  for (const session of sessions) {
-    if (session.title) {
-      sessionTitleCache.set(session.title, session.id);
-    }
-  }
-  sessionListCacheLastLoaded = now;
-
-  return sessionTitleCache.get(sessionTarget.title) ?? null;
-}
